fix: Lizard에 현재 몸무게 추가 및 데이터 타입 float으로 변경

+++ kiiun/src/main/java/com/cx_project/kiiun/domain/growth/entity/Growth.java
@@ -26,7 +26,7 @@ public class Growth {
     private Long id;
 
     @Column(name = "weight")
-    private Integer weight;
+    private float weight;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "lizard_id", nullable = false)
